Move methods from BaseConverter to AbstractConverter.
