+++ src/dodecahedron/units_of_work/sessioned_unit_of_work.py
@@ -18,7 +18,6 @@ from typing import TypeVar
 # Local Imports
 from .abstract_unit_of_work import AbstractUnitOfWork
 
-
 # Custom types
 T = TypeVar("T")
